Create thermistor-controlled-speaker-and-neopixel.py
I don't know what the values of my thermistor are so I didn't dial in the settings at all. 
It reads a temperature, the degrees I translate it into might be highly inaccurate - I only somewhat calibrated the base level to match a regular thermometer. 

Goes from a cool blue to a warm orange to a red hot, with the red hot triggering the speaker to boot as an alarm.